FLS/baseline_sonar_yolo/requirements.txt
ultralytics>=8.2.0
opencv-python>=4.9.0
pandas>=2.2.0
numpy>=1.26.0
PyYAML>=6.0.1
matplotlib>=3.8.0
